Back-end/requirements.txt
django
djangorestframework
Pillow
django-cors-headers
djangorestframework-simplejwt
python-decouple
gunicorn
django-on-heroku
